requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==2.1.0
Django==4.2.23
django-allauth==65.9.0
django-anymail==10.3
django-browser-reload==1.12.0
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-payments==2.0.0
django-phonenumber-field==7.3.0
django-ratelimit==4.1.0
django-storages==1.14.2
django-tailwind==3.6.0
django-unfold==0.34.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
environ==1.0
google-api-core==2.25.1
google-api-python-client==2.126.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==2.12.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
packaging==25.0
phonenumberslite==9.0.7
pillow==10.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.3
pytz==2023.3.post1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.6.0
